+++ src/ui/components/Switch.svelte
@@ -38,14 +38,14 @@
   }
 
   input[type='checkbox']:disabled ~ .switch .track {
-    @apply bg-gray-500 cursor-default;
+    @apply bg-gray-300 cursor-not-allowed;
   }
   input[type='checkbox']:disabled:active + .switch .track {
     @apply shadow-none;
   }
 
   input[type='checkbox']:disabled ~ .switch .thumb {
-    @apply bg-gray-100 border-gray-500;
+    @apply bg-gray-100 border-gray-300 cursor-not-allowed;
   }
 
   input[type='checkbox']:focus + .switch .track,